Extract boilerplate localization classes from example's main
